+++ reset.sh
@@ -83,7 +83,9 @@ echo -e "${green}server successfully stopped!${nocolor}"
 
 # create backup
 echo -e "${blue}backing up...${nocolor}"
-rm -r ${backupdirectory}/cached/reset-*
+if [ -d "${backupdirectory}/cached/reset-*" ]; then
+	rm -r ${backupdirectory}/cached/reset-*
+fi
 cp -r ${serverdirectory}/world ${backupdirectory}/cached/reset-${newdaily}
 
 # check if safety backup exists
